This project will be designing the application that will help n the managing of flights and passengers. The application will have six forms with user interfaces that users will be able to use to manage the required information when making bookings for the flights.
There are log in, home, passenger, flights, about and exit forms. When the application is launched the user will be asked for log in details.  The user will put in credentials, if the credentials are wrong, the error message will be displayed else the user will be taken to the home page where will have links to other pages of the application. The log in page has a cancel button as well that can be used to exit the application.

The main forms are passengers and flights. The passenger form is used to capture the information about the passenger who is trying to book for a flight. On this for the passenger will need to provide the name, surname, ID and date of birth. These will be linked to the flight that the passenger is trying to book.
There will also be a back button that will be used to go back to the home page in case the user has changed the mind or would like to make some changes.
The next is the flight form, on here information about flights are recorded. The information which include the departure times, flight name and number of seats will be provide to ensure that correct passengers book the correct flight and their availabilities.

Signed-off-by: Manfred Brown <48109657@unisa.ac.za>